chore: fix lint and test
