+++ tests/mocks/shpParser.ts
@@ -3,9 +3,9 @@ import { ShpParser } from '../../src/layerCreator/models/shpParser';
 const parseMock = jest.fn();
 const toGeoJsonMock = jest.fn();
 
-const ShpParserMock = ({
-    parse: parseMock,
-    toGeoJson: toGeoJsonMock
+const shpParserMock = ({
+  parse: parseMock,
+  toGeoJson: toGeoJsonMock,
 } as unknown) as ShpParser;
 
-export { ShpParserMock, parseMock, toGeoJsonMock };
+export { shpParserMock, parseMock, toGeoJsonMock };
